Move edge logout util to edge utils
