+++ client/src/util/edge.ts
@@ -1,5 +1,6 @@
 import { EdgeAccount } from 'edge-core-js';
 import { WalletKeys } from '../types';
+import { sleep } from '../utils';
 
 export const getWalletKeys = async (
   account: EdgeAccount,
@@ -21,4 +22,13 @@ export const getWalletKeys = async (
   }
 
   return keys;
+};
+
+export const waitForEdgeAccountStop = async (edgeAccount: EdgeAccount) => {
+  try {
+    edgeAccount && edgeAccount.loggedIn && (await edgeAccount.logout());
+    await sleep(3000); // todo: added for testnet env because edge fio plugin rewriting base url after pin confirmation. Need to figure out how to be sure that fio engine is killed or fix transaction global base url const in FIOSDK
+  } catch (e) {
+    //
+  }
 };
